pulled header into component, panels and header and main classnames into Layout.tsx
