+++ src/pages/_app.tsx
@@ -1,24 +1,12 @@
-import { type AppType } from "next/app";
-import { Inter } from "next/font/google";
-import { cn } from "@/lib/utils";
-
 import "@/styles/globals.css";
-
-const inter = Inter({
-  subsets: ["latin"],
-  variable: "--font-sans",
-});
+import { type AppType } from "next/app";
+import Layout from "@/components/Layout";
 
 const MyApp: AppType = ({ Component, pageProps }) => {
   return (
-    <main
-      className={cn(
-        "h-[100vh] bg-background font-sans antialiased",
-        inter.variable
-      )}
-    >
+    <Layout>
       <Component {...pageProps} />
-    </main>
+    </Layout>
   );
 };
